+++ includes/system/class-cache.php
@@ -10,6 +10,8 @@
 
 namespace Decalog\System;
 
+use Decalog\System\Conversion;
+
 /**
  * The class responsible to handle cache management.
  *
@@ -28,30 +30,6 @@ class Cache {
 	private static $pool_name = DECALOG_SLUG;
 
 	/**
-	 * Differentiates cache items by blogs.
-	 *
-	 * @since  1.0.0
-	 * @var    boolean    $blog_aware    Is the item id must contain the blog id?
-	 */
-	private static $blog_aware = false;
-
-	/**
-	 * Differentiates cache items by current locale.
-	 *
-	 * @since  1.0.0
-	 * @var    boolean    $blog_aware    Is the item id must contain the locale id?
-	 */
-	private static $locale_aware = false;
-
-	/**
-	 * Differentiates cache items by current user.
-	 *
-	 * @since  1.0.0
-	 * @var    boolean    $blog_aware    Is the item id must contain the user id?
-	 */
-	private static $user_aware = false;
-
-	/**
 	 * Available TTLs.
 	 *
 	 * @since  1.0.0
@@ -66,6 +44,38 @@ class Cache {
 	 * @var    integer    $default_ttl    The default TTL in seconds.
 	 */
 	private static $default_ttl = 3600;
+
+	/**
+	 * Is APCu available.
+	 *
+	 * @since  1.0.0
+	 * @var    boolean    $apcu_available    Is APCu available.
+	 */
+	private static $apcu_available = false;
+
+	/**
+	 * Hits values.
+	 *
+	 * @since  1.0.0
+	 * @var    array    $hit    Hits values.
+	 */
+	private static $hit = [];
+
+	/**
+	 * Miss values.
+	 *
+	 * @since  1.0.0
+	 * @var    array    $miss    Miss values.
+	 */
+	private static $miss = [];
+
+	/**
+	 * Current (temporary) values.
+	 *
+	 * @since  1.0.0
+	 * @var    array    $current    Current (temporary) values.
+	 */
+	private static $current = [];
 
 	/**
 	 * Initializes the class and set its properties.
@@ -88,6 +98,11 @@ class Cache {
 			'diagnosis'         => HOUR_IN_SECONDS,
 			'plugin-statistics' => DAY_IN_SECONDS,
 		];
+		if ( wp_using_ext_object_cache() ) {
+			wp_cache_add_global_groups( self::$pool_name );
+		}
+		self::$apcu_available = function_exists( 'apcu_delete' ) && function_exists( 'apcu_fetch' ) && function_exists( 'apcu_store' );
+		add_action( 'shutdown', [ 'Decalog\System\Cache', 'log_debug' ], 10, 0 );
 	}
 
 	/**
@@ -95,30 +110,57 @@ class Cache {
 	 *
 	 * @since 1.0.0
 	 */
-	public static function id( $args, $path = '/Data' ) {
-		return $path . '/Raw/' . md5( (string) $args );
+	public static function id( $args, $path = 'data/' ) {
+		if ( '/' === $path[0] ) {
+			$path = substr( $path, 1 );
+		}
+		if ( '/' !== $path[ strlen( $path ) - 1 ] ) {
+			$path = $path . '/';
+		}
+		return $path . md5( (string) $args );
 	}
 
 	/**
 	 * Full item name.
 	 *
-	 * @param  string $item_name Item name. Expected to not be SQL-escaped.
+	 * @param  string  $item_name Item name. Expected to not be SQL-escaped.
+	 * @param  boolean $blog_aware   Optional. Has the name must take care of blog.
+	 * @param  boolean $locale_aware Optional. Has the name must take care of locale.
+	 * @param  boolean $user_aware   Optional. Has the name must take care of user.
 	 * @return string The full item name.
 	 * @since  1.0.0
 	 */
-	private static function full_item_name( $item_name ) {
-		$name = self::$pool_name . '/';
-		if ( self::$blog_aware ) {
+	private static function full_item_name( $item_name, $blog_aware = false, $locale_aware = false, $user_aware = false ) {
+		$name = '';
+		if ( $blog_aware ) {
 			$name .= (string) get_current_blog_id() . '/';
 		}
-		if ( self::$locale_aware ) {
+		if ( $locale_aware ) {
 			$name .= (string) L10n::get_display_locale() . '/';
 		}
-		if ( self::$user_aware ) {
+		if ( $user_aware ) {
 			$name .= (string) User::get_current_user_id() . '/';
 		}
 		$name .= $item_name;
-		return substr( trim( $name ), 0, 172 );
+		return substr( trim( $name ), 0, 172 - strlen( self::$pool_name ) );
+	}
+
+	/**
+	 * Normalized item name.
+	 *
+	 * @param  string $item_name Item name. Expected to not be SQL-escaped.
+	 * @return string The normalized item name.
+	 * @since  1.0.0
+	 */
+	private static function normalized_item_name( $item_name ) {
+		if ( '/' === $item_name[0] ) {
+			$item_name = substr( $item_name, 1 );
+		}
+		while ( 0 !== substr_count( $item_name, '//' ) ) {
+			$item_name = str_replace( '//', '/', $item_name );
+		}
+		$item_name = str_replace( '/', '_', $item_name );
+		return strtolower( $item_name );
 	}
 
 	/**
@@ -132,11 +174,27 @@ class Cache {
 	 * @since  1.0.0
 	 */
 	private static function get_for_full_name( $item_name ) {
-		while ( 0 !== substr_count( $item_name, '//' ) ) {
-			$item_name = str_replace( '//', '/', $item_name );
+		$chrono    = microtime( true );
+		$item_name = self::normalized_item_name( $item_name );
+		$found     = false;
+		if ( wp_using_ext_object_cache() ) {
+			$result = wp_cache_get( $item_name, self::$pool_name, false, $found );
+		} elseif ( self::$apcu_available ) {
+			$result = apcu_fetch( self::$pool_name . '_' . $item_name, $found );
+		} else {
+			$result = get_transient( self::$pool_name . '_' . $item_name );
+			$found  = false !== $result;
 		}
-		$item_name = str_replace( '/', '_', $item_name );
-		return get_transient( $item_name );
+		if ( $found ) {
+			self::$hit[] = [
+				'time' => microtime( true ) - $chrono,
+				'size' => strlen( serialize( $result ) ),
+			];
+			return $result;
+		} else {
+			self::$current[ $item_name ] = $chrono;
+			return null;
+		}
 	}
 
 	/**
@@ -150,7 +208,7 @@ class Cache {
 	 * @since  1.0.0
 	 */
 	public static function get_global( $item_name ) {
-		return self::get_for_full_name( self::$pool_name . '/' . $item_name );
+		return self::get_for_full_name( self::full_item_name( $item_name ) );
 	}
 
 	/**
@@ -159,12 +217,15 @@ class Cache {
 	 * If the item does not exist, does not have a value, or has expired,
 	 * then the return value will be false.
 	 *
-	 * @param  string $item_name Item name. Expected to not be SQL-escaped.
+	 * @param  string  $item_name Item name. Expected to not be SQL-escaped.
+	 * @param  boolean $blog_aware   Optional. Has the name must take care of blog.
+	 * @param  boolean $locale_aware Optional. Has the name must take care of locale.
+	 * @param  boolean $user_aware   Optional. Has the name must take care of user.
 	 * @return mixed Value of item.
 	 * @since  1.0.0
 	 */
-	public static function get( $item_name ) {
-		return self::get_for_full_name( self::full_item_name( $item_name ) );
+	public static function get( $item_name, $blog_aware = false, $locale_aware = false, $user_aware = false ) {
+		return self::get_for_full_name( self::full_item_name( $item_name, $blog_aware, $locale_aware, $user_aware ) );
 	}
 
 	/**
@@ -181,19 +242,29 @@ class Cache {
 	 * @since  1.0.0
 	 */
 	private static function set_for_full_name( $item_name, $value, $ttl = 'default' ) {
-		while ( 0 !== substr_count( $item_name, '//' ) ) {
-			$item_name = str_replace( '//', '/', $item_name );
-		}
-		$item_name  = str_replace( '/', '_', $item_name );
+		$item_name  = self::normalized_item_name( $item_name );
 		$expiration = self::$default_ttl;
 		if ( array_key_exists( $ttl, self::$ttls ) ) {
 			$expiration = self::$ttls[ $ttl ];
 		}
 		if ( $expiration >= 0 ) {
-			return set_transient( $item_name, $value, $expiration );
+			if ( wp_using_ext_object_cache() ) {
+				$result = wp_cache_set( $item_name, $value, self::$pool_name, (int) $expiration );
+			} elseif ( self::$apcu_available ) {
+				$result = apcu_store( self::$pool_name . '_' . $item_name, $value, $expiration );
+			} else {
+				$result = set_transient( self::$pool_name . '_' . $item_name, $value, $expiration );
+			}
+			if ( array_key_exists( $item_name, self::$current ) ) {
+				self::$miss[] = [
+					'time' => microtime( true ) - self::$current[ $item_name ],
+					'size' => strlen( serialize( $result ) ),
+				];
+			}
 		} else {
-			return false;
+			$result = false;
 		}
+		return $result;
 	}
 
 	/**
@@ -210,7 +281,7 @@ class Cache {
 	 * @since  1.0.0
 	 */
 	public static function set_global( $item_name, $value, $ttl = 'default' ) {
-		return self::set_for_full_name( self::$pool_name . '/' . $item_name, $value, $ttl );
+		return self::set_for_full_name( self::full_item_name( $item_name ), $value, $ttl );
 	}
 
 	/**
@@ -219,34 +290,49 @@ class Cache {
 	 * You do not need to serialize values. If the value needs to be serialized, then
 	 * it will be serialized before it is set.
 	 *
-	 * @param  string $item_name Item name. Expected to not be SQL-escaped.
-	 * @param  mixed  $value     Item value. Must be serializable if non-scalar.
-	 *                           Expected to not be SQL-escaped.
-	 * @param  string $ttl       Optional. The previously defined ttl @see self::init().
+	 * @param  string  $item_name    Item name. Expected to not be SQL-escaped.
+	 * @param  mixed   $value        Item value. Must be serializable if non-scalar.
+	 *                               Expected to not be SQL-escaped.
+	 * @param  string  $ttl          Optional. The previously defined ttl @see self::init().
+	 * @param  boolean $blog_aware   Optional. Has the name must take care of blog.
+	 * @param  boolean $locale_aware Optional. Has the name must take care of locale.
+	 * @param  boolean $user_aware   Optional. Has the name must take care of user.
 	 * @return bool False if value was not set and true if value was set.
 	 * @since  1.0.0
 	 */
-	public static function set( $item_name, $value, $ttl = 'default' ) {
-		return self::set_for_full_name( self::full_item_name( $item_name ), $value, $ttl );
+	public static function set( $item_name, $value, $ttl = 'default', $blog_aware = false, $locale_aware = false, $user_aware = false ) {
+		return self::set_for_full_name( self::full_item_name( $item_name, $blog_aware, $locale_aware, $user_aware ), $value, $ttl );
 	}
 
 	/**
 	 * Delete the value of a fully named cache item.
 	 *
-	 * This function accepts generic car "*".
+	 * This function accepts generic car "*" for transients.
 	 *
 	 * @param  string $item_name Item name. Expected to not be SQL-escaped.
 	 * @return integer Number of deleted items.
 	 * @since  1.0.0
 	 */
 	private static function delete_for_ful_name( $item_name ) {
-		global $wpdb;
-		while ( 0 !== substr_count( $item_name, '//' ) ) {
-			$item_name = str_replace( '//', '/', $item_name );
-		}
-		$item_name = str_replace( '/', '_', $item_name );
+		$item_name = self::normalized_item_name( $item_name );
 		$result    = 0;
-		if ( strlen( $item_name ) - 1 === strpos( $item_name, '/*' ) && '/' === $item_name[0] ) {
+		if ( wp_using_ext_object_cache() ) {
+			if ( strlen( $item_name ) - 1 === strpos( $item_name, '_*' ) ) {
+				return false;
+			} else {
+				return wp_cache_delete( $item_name, self::$pool_name );
+			}
+		}
+		if ( self::$apcu_available ) {
+			if ( strlen( $item_name ) - 1 === strpos( $item_name, '_*' ) ) {
+				return false;
+			} else {
+				return apcu_delete( $item_name );
+			}
+		}
+		global $wpdb;
+		$item_name = self::$pool_name . '_' . $item_name;
+		if ( strlen( $item_name ) - 1 === strpos( $item_name, '_*' ) ) {
 			// phpcs:ignore
 			$delete = $wpdb->get_col( "SELECT option_name FROM {$wpdb->options} WHERE option_name = '_transient_timeout_" . str_replace( '_*', '', $item_name ) . "' OR option_name LIKE '_transient_timeout_" . str_replace( '_*', '_%', $item_name ) . "';" );
 		} else {
@@ -265,27 +351,91 @@ class Cache {
 	/**
 	 * Delete the value of a global cache item.
 	 *
-	 * This function accepts generic car "*".
+	 * This function accepts generic car "*" for transients.
 	 *
 	 * @param  string $item_name Item name. Expected to not be SQL-escaped.
 	 * @return integer Number of deleted items.
 	 * @since  1.0.0
 	 */
 	public static function delete_global( $item_name ) {
-		return self::delete_for_ful_name( self::$pool_name . '/' . $item_name );
+		return self::delete_for_ful_name( self::full_item_name( $item_name ) );
 	}
 
 	/**
 	 * Delete the value of a standard cache item.
 	 *
-	 * This function accepts generic car "*".
+	 * This function accepts generic car "*" for transients.
 	 *
-	 * @param  string $item_name Item name. Expected to not be SQL-escaped.
+	 * @param  string  $item_name Item name. Expected to not be SQL-escaped.
+	 * @param  boolean $blog_aware   Optional. Has the name must take care of blog.
+	 * @param  boolean $locale_aware Optional. Has the name must take care of locale.
+	 * @param  boolean $user_aware   Optional. Has the name must take care of user.
 	 * @return integer Number of deleted items.
 	 * @since  1.0.0
 	 */
-	public static function delete( $item_name ) {
-		return self::delete_for_ful_name( self::full_item_name( $item_name ) );
+	public static function delete( $item_name, $blog_aware = false, $locale_aware = false, $user_aware = false ) {
+		return self::delete_for_ful_name( self::full_item_name( $item_name, $blog_aware, $locale_aware, $user_aware ) );
+	}
+
+	/**
+	 * Get cache analytics.
+	 *
+	 * @return array The cache analytics.
+	 * @since  1.0.0
+	 */
+	public static function get_analytics() {
+		$result    = [];
+		$hit_time  = 0;
+		$hit_count = count( self::$hit );
+		$hit_size  = 0;
+		if ( 0 < $hit_count ) {
+			foreach ( self::$hit as $h ) {
+				$hit_time = $hit_time + $h['time'];
+				$hit_size = $hit_size + $h['size'];
+			}
+			$hit_time = $hit_time / $hit_count;
+			$hit_size = $hit_size / $hit_count;
+		}
+		$result['hit']['count'] = $hit_count;
+		$result['hit']['time'] = $hit_time;
+		$result['hit']['size'] = $hit_size;
+		$miss_time  = 0;
+		$miss_count = count( self::$miss );
+		$miss_size  = 0;
+		if ( 0 < $miss_count ) {
+			foreach ( self::$miss as $h ) {
+				$miss_time = $miss_time + $h['time'];
+				$miss_size = $miss_size + $h['size'];
+			}
+			$miss_time = $miss_time / $miss_count;
+			$miss_size = $miss_size / $miss_count;
+		}
+		$result['miss']['count'] = $miss_count;
+		$result['miss']['time'] = $miss_time;
+		$result['miss']['size'] = $miss_size;
+		if ( wp_using_ext_object_cache() ) {
+			$result['type'] = 'object_cache';
+		} elseif ( self::$apcu_available ) {
+			$result['type'] = 'apcu';
+		} else {
+			$result['type'] = 'db_transient';
+		}
+		return $result;
+	}
+
+	/**
+	 * Logs the cache analytics.
+	 *
+	 * @since  1.0.0
+	 */
+	public static function log_debug() {
+		$analytics = self::get_analytics();
+		$log       = '[' . $analytics['type'] . ']';
+		$log      .= '   Hit count: ' . $analytics['hit']['count'] . '   Hit time: ' . round($analytics['hit']['time'] * 1000, 3) . 'ms   Hit size: ' . Conversion::data_shorten( (int) $analytics['hit']['size'] );
+		$log      .= '   Miss count: ' . $analytics['miss']['count'] . '   Miss time: ' . round($analytics['miss']['time'] * 1000, 3) . 'ms   Miss size: ' . Conversion::data_shorten( (int) $analytics['miss']['size'] );
+		if ( 0 !== (int) $analytics['hit']['count'] || 0 !== (int) $analytics['miss']['count'] ) {
+			Logger::debug( $log );
+		}
 	}
 
 }